Move test functions to their own module
Also, move `dandiset_id` attribute from `Asset` to `TestCase`
