+++ code/src/healthstatus/__main__.py
@@ -12,10 +12,11 @@ from time import sleep
 import anyio
 import click
 import requests
-from .checker import TEST_NAMES, HealthStatus
+from .checker import HealthStatus
 from .config import MATNWB_INSTALL_DIR
 from .core import Outcome, log
 from .reporter import DandisetStatus, TestSummary
+from .tests import TESTS
 
 
 @click.group()
@@ -115,7 +116,7 @@ def report() -> None:
         Outcome.TIMEOUT: 0,
     }
     all_statuses = []
-    test_summaries = {tn: TestSummary(tn) for tn in TEST_NAMES}
+    test_summaries = {tn: TestSummary(tn) for tn in TESTS.keys()}
     for p in Path("results").iterdir():
         if re.fullmatch(r"\d{6,}", p.name) and p.is_dir():
             status = DandisetStatus.from_file(p.name, p / "status.yaml")
@@ -126,7 +127,7 @@ def report() -> None:
             dandiset_qtys[Outcome.PASS] += bool(not failed and not timedout and passed)
             dandiset_qtys[Outcome.FAIL] += bool(failed)
             dandiset_qtys[Outcome.TIMEOUT] += bool(timedout)
-            for tn in TEST_NAMES:
+            for tn in TESTS.keys():
                 test_summaries[tn].register(p.name, status.test_counts(tn))
             all_statuses.append(status)
     with open("README.md", "w") as fp:
@@ -139,11 +140,11 @@ def report() -> None:
             file=fp,
         )
         print("| --- | --- | --- | --- |", file=fp)
-        for tn in TEST_NAMES:
+        for tn in TESTS.keys():
             print(test_summaries[tn].as_row(), file=fp)
         print(file=fp)
-        print("| Dandiset | " + " | ".join(TEST_NAMES) + " | Untested |", file=fp)
-        print("| --- | " + " | ".join("---" for _ in TEST_NAMES) + " | --- |", file=fp)
+        print("| Dandiset | " + " | ".join(TESTS.keys()) + " | Untested |", file=fp)
+        print("| --- | " + " | ".join("---" for _ in TESTS) + " | --- |", file=fp)
         for s in sorted(all_statuses, key=attrgetter("identifier")):
             print(s.as_row(), file=fp)
